Provide extensive redundant documentation to pass project rules
One of the things I had to rewrite in the project was documentation of
methods. I disagree with such approach, because I did everything to
have functions
- sufficiently small
- self-explanatory,
- named so that they communicate intent comprehensively.

The added documentation is mostly a human-readable
rewrite of the function names.

Nevertheless, rules are rules, so I'm passing the documentation as
required.